fix for new GCC versions (Fedora Core 6)

+++ code/windows_stub/stubs.cpp
@@ -1,13 +1,16 @@
 
 /*
  * $Logfile: $
- * $Revision: 2.32.2.1 $
- * $Date: 2006-09-08 06:14:44 $
+ * $Revision: 2.32.2.2 $
+ * $Date: 2006-12-07 18:26:48 $
  * $Author: taylor $
  *
  * OS-dependent functions.
  *
  * $Log: not supported by cvs2svn $
+ * Revision 2.32.2.1  2006/09/08 06:14:44  taylor
+ * fix things that strict compiling balked at (from compiling with -ansi and -pedantic)
+ *
  * Revision 2.32  2006/05/27 16:39:40  taylor
  * remove non-Windows timeSetEvent() functions
  * make PDWORD and LPDWORD int instead of long (64-bit issue) (thanks Spike)
@@ -381,7 +384,7 @@ bool QueryPerformanceCounter(LARGE_INTEGER *pcount)
 
 	gettimeofday(&timer_now, NULL);
 
-	pcount->QuadPart = (long long)timer_now.tv_usec;
+	pcount->QuadPart = (longlong)timer_now.tv_usec;
 	
 	return 1;
 }
